refactor: introduce library crate and move shared types from main
Introduce a library crate entry point (lib.rs) and move shared
application types out of main.rs.

This begins separating the CLI entrypoint from reusable library code,
which will allow the RetroMount domain model, translators, and FUSE
integration to live in the library crate while keeping main.rs thin.

No functional behaviour changes.
